Construct a graph object's mixing matrix.

Using a categorical vertex attribute, construct a matrix depicting network mixing patterns.

as_mixing_matrix(g, row_attr, col_attr = NULL, direction = c("all",
  "out", "in"))

Arguments

g
<igraph>, <network>

row_attr
Either the name of a node/vertex attribute or vector of same length as the number of nodes in g.

col_attr
If not provided, the same as row_attr. Otherwise, a vector of the same length as the number of nodes in g.

direction
<chr> The direction of ties to take into a account.

"all", "out", or "in"

Examples

as_mixing_matrix(jemmah_islamiyah, row_attr = "role")

#> 5 x 5 Matrix of class "dtrMatrix"
#>                      
#>                       command team operation assistant bomb maker
#>   command team                   6                  16         30
#>   operation assistant            .                   2         10
#>   bomb maker                     .                   .         20
#>   suicide bomber                 .                   .          .
#>   Team Lima                      .                   .          .
#>                      
#>                       suicide bomber Team Lima
#>   command team                     2         8
#>   operation assistant              2         0
#>   bomb maker                      10         0
#>   suicide bomber                   0         8
#>   Team Lima                        .        12

data("sampson", package = "ergm")
as_mixing_matrix(samplike, row_attr = "vertex.names", col_attr = "group",
                 direction = "all")

#> 18 x 3 Matrix of class "dgeMatrix"
#>              
#>               Turks Outcasts Loyal
#>   John Bosco      9        3     5
#>   Gregory        11        3     1
#>   Basil           3        5     0
#>   Peter           0        0     9
#>   Bonaventure     3        2     8
#>   Berthold        1        0     5
#>   Mark            8        2     1
#>   Victor          4        0     7
#>   Ambrose         2        0     6
#>   Romauld         1        1     6
#>   Louis           3        0     5
#>   Winfrid        10        0     1
#>   Amand           1        4     3
#>   Hugh            8        0     3
#>   Boniface        8        0     1
#>   Albert          6        0     2
#>   Elias           1        5     0
#>   Simplicius      3        6     0

as_mixing_matrix(samplike, row_attr = "vertex.names", col_attr = "group",
                 direction = "out")

#> 18 x 3 Matrix of class "dgeMatrix"
#>              
#>               Turks Outcasts Loyal
#>   John Bosco      3        1     2
#>   Gregory         5        0     0
#>   Basil           2        3     0
#>   Peter           0        0     4
#>   Bonaventure     1        1     3
#>   Berthold        1        0     3
#>   Mark            4        0     1
#>   Victor          2        0     4
#>   Ambrose         2        0     2
#>   Romauld         1        1     4
#>   Louis           2        0     3
#>   Winfrid         4        0     0
#>   Amand           1        1     1
#>   Hugh            4        0     1
#>   Boniface        5        0     1
#>   Albert          5        0     0
#>   Elias           1        3     0
#>   Simplicius      3        3     0

as_mixing_matrix(samplike, row_attr = "vertex.names", col_attr = "group",
                 direction = "in")

#> 18 x 3 Matrix of class "dgeMatrix"
#>              
#>               Turks Outcasts Loyal
#>   John Bosco      6        2     3
#>   Gregory         6        3     1
#>   Basil           1        2     0
#>   Peter           0        0     5
#>   Bonaventure     2        1     5
#>   Berthold        0        0     2
#>   Mark            4        2     0
#>   Victor          2        0     3
#>   Ambrose         0        0     4
#>   Romauld         0        0     2
#>   Louis           1        0     2
#>   Winfrid         6        0     1
#>   Amand           0        3     2
#>   Hugh            4        0     2
#>   Boniface        3        0     0
#>   Albert          1        0     2
#>   Elias           0        2     0
#>   Simplicius      0        3     0