Krackhardt's E-I Index

Given a categorical vertex attribute describing mutually exclusive groups, the E-I index represents a ratio of external to internal ties.

ei_index(g, node_attr_name, scope = c("global", "group", "node"))

# S3 method for network
ei_index(g, node_attr_name, scope = c("global",
  "group", "node"))

# S3 method for igraph
ei_index(g, node_attr_name, scope = c("global", "group",
  "node"))

Arguments

g

<igraph>

<network>, if {intergraph} is installed.

node_attr_name

<chr>

The name of a node/vertex attribute in g.

scope

<chr>

The target scope of the resulting EI-Index

"global", "group", or "node"

Details

$$E\mbox{-}I~Index = \frac{EL-IL}{EL+IL}$$

References

Krackhardt, David, and Robert N. Stern. "Informal Networks and Organizational Crises: An Experimental Simulation." Social Psychology Quarterly 51, no. 2 (1988): 123-40. http://www.jstor.org/stable/2786835.

Examples

ei_index(jemmah_islamiyah, node_attr_name = "role")

#> [1] 0.3650794

ei_index(jemmah_islamiyah, node_attr_name = "role", scope = "group")

#>             attribute external_ties internal_ties   ei_index
#> 1          bomb maker             5            10 -0.3333333
#> 2        command team            23             3  0.7692308
#> 3 operation assistant            10             1  0.8181818
#> 4      suicide bomber             5             0  1.0000000
#> 5           Team Lima             0             6 -1.0000000

ei_index(jemmah_islamiyah, node_attr_name = "role", scope = "node")

#>        node           attribute external_ties internal_ties   ei_index
#> 1    MUKLAS        command team             7             2  0.5555556
#> 2    AMROZI operation assistant             3             1  0.5000000
#> 3     IMRON operation assistant             9             0  1.0000000
#> 4   SAMUDRA        command team            13             2  0.7333333
#> 5  DULMATIN          bomb maker             5             4  0.1111111
#> 6     IDRIS        command team             8             2  0.6000000
#> 7   MUBAROK operation assistant             2             1  0.3333333
#> 8   AZAHARI          bomb maker             5             4  0.1111111
#> 9     GHONI          bomb maker             5             4  0.1111111
#> 10  ARNASAN      suicide bomber             5             0  1.0000000
#> 11     RAUF           Team Lima             2             3 -0.2000000
#> 12  OCTAVIA           Team Lima             2             3 -0.2000000
#> 13  HIDAYAT           Team Lima             2             3 -0.2000000
#> 14  JUNAEDI           Team Lima             2             3 -0.2000000
#> 15    PATEK          bomb maker             5             4  0.1111111
#> 16     FERI      suicide bomber             6             0  1.0000000
#> 17   SARIJO          bomb maker             5             4  0.1111111