Krackhardt's E-I Index

Given a categorical vertex attribute describing mutually exclusive groups, the E-I index represents a ratio of external to internal ties.

ei_index(g, node_attr_name, scope = c("global", "group", "node"), ...)

Arguments

g
<igraph>, <network>

node_attr_name
<chr> The name of a node/vertex attribute in g.

scope
<chr> The target scope of the resulting EI-Index

"global", "group", or "node"

...
Arguments passed to or from other methods.

Details

$$E\mbox{-}I~Index = \frac{EL-IL}{EL+IL}$$

\(EL\): external ties

\(IL\): internal ties

References

Krackhardt, David, and Robert N. Stern. "Informal Networks and Organizational Crises: An Experimental Simulation." Social Psychology Quarterly 51, no. 2 (1988): 123-40. http://www.jstor.org/stable/2786835.

Examples

# <igraph>
ei_index(jemmah_islamiyah, node_attr_name = "role")

#> [1] 0.3650794

ei_index(jemmah_islamiyah, node_attr_name = "role", scope = "group")

#>        command team operation assistant          bomb maker      suicide bomber 
#>           0.8064516           0.7142857          -0.3333333           1.0000000 
#>           Team Lima 
#>          -1.0000000 

ei_index(jemmah_islamiyah, node_attr_name = "role", scope = "node")

#>     MUKLAS     AMROZI      IMRON    SAMUDRA   DULMATIN      IDRIS    MUBAROK 
#>  0.5555556  0.5000000  1.0000000  0.7333333  0.1111111  0.6000000  0.3333333 
#>    AZAHARI      GHONI    ARNASAN       RAUF    OCTAVIA    HIDAYAT    JUNAEDI 
#>  0.1111111  0.1111111  1.0000000 -0.2000000 -0.2000000 -0.2000000 -0.2000000 
#>      PATEK       FERI     SARIJO 
#>  0.1111111  1.0000000  0.1111111 

# <network>
data("sampson", package = "ergm")
ei_index(samplike, node_attr_name = "group")

#> [1] -0.4318182

ei_index(samplike, node_attr_name = "group", scope = "group")

#>      Turks   Outcasts      Loyal 
#> -0.6666667 -0.8181818 -1.0000000 

ei_index(samplike, node_attr_name = "group", scope = "node")

#>  John Bosco     Gregory       Basil       Peter Bonaventure    Berthold 
#> -0.05882353 -0.46666667 -0.25000000 -1.00000000 -0.23076923 -0.66666667 
#>        Mark      Victor     Ambrose     Romauld       Louis     Winfrid 
#> -0.45454545 -0.27272727 -0.50000000 -0.50000000 -0.25000000 -0.81818182 
#>       Amand        Hugh    Boniface      Albert       Elias  Simplicius 
#>  0.00000000 -0.45454545 -0.77777778 -0.50000000 -0.66666667 -0.33333333